🎨 Improve the code
Improve dnf-upgrade, implement aur.sh and pacman.sh, refactor the code and fix some bad code

+++ kill.sh
@@ -7,28 +7,31 @@
 #	License:		This file is licensed under the LGPLv3.						#
 #################################################################################
 
+# Kill all the instance of a program
 function kill-kill {
+	# the parameter $1 is the name of the program that you want kill
 	killall -s KILL -v -I $1
 }
 
+# Kill all the PID's into the file processes_to_kill.txt
 function kills {
-	path=$(find $HOME -type f -regex ".*processes_to_kill\.txt")	# retrieve the path of the file that contains the PIDs of the processes that must be killed
+	path=$(find $HOME -type f -regex ".*processes_to_kill\.txt" 2> /dev/null)	# retrieve the path of the file that contains the PIDs of the processes that must be killed
 
-	if [ -e $path ] && [ -f $path ]									# if that checks if the file exists
+	if [ $? -eq 0 ] && [ -e $path ] && [ -f $path ]								# if that checks if the file exists
 	then
 		buffer=''
 
-		for i in $(cat $path)										# create a list with the PIDs into the file
+		for i in $(cat $path)													# create a list with the PIDs into the file
 		do
 			buffer="${buffer}${i} "
 		done
 
-		rm -rf $path												# delete permanently the PIDs
+		rm -rf $path															# delete permanently the PIDs
 
 		let i=${#buffer} - 1
 		buffer=${buffer:$i}
 
-		kill -s KILL $buffer > /dev/null							# kill the processes
+		kill -s KILL $buffer > /dev/null										# kill the processes
 	else
 		echo -e "There aren\'t process in background."
 	fi
